PE-0119: runs faster but answer doesn't seem right.

+++ Demo.ProjectEuler/Demo.ProjectEuler.Core/NumberUtil.cs
@@ -5,11 +5,11 @@ namespace Demo.ProjectEuler.Core
 {
 	public class NumberUtil
 	{
-		public IEnumerable<long> ToReverseSequence(BigInteger number)
+		public IEnumerable<int> ToReverseSequence(BigInteger number)
 		{
 			while (number > 0)
 			{
-				long digit = (long) number % 10;
+				int digit = (int) (number % 10);
 				yield return digit;
 
 				number /= 10;
